# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.10.9-SNAPSHOT"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /apis/{apiId}/pages/{pageId}/content:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
      - $ref: '#/components/parameters/pageIdParam'
    get:
      tags:
        - Api
      summary: Get the content of an API page.
      description: |
        Get the content of a specific API documentation page.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getPageContentByApiIdAndPageId
      security: []
      responses:
        200:
          description: The content of the api documentation page.
          content:
            text/plain:
              schema:
                type: string
        404:
          description: API or Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /pages/{pageId}:
    parameters:
      - $ref: '#/components/parameters/pageIdParam'
    get:
      tags:
        - Portal
      parameters:
        - $ref: '#/components/parameters/pageIncludeParam'
      summary: Get a portal page
      description: |
        Get a specific portal documentation page.
      operationId: getPageByPageId
      security: []
      responses:
        200:
          $ref: '#/components/responses/PageSuccess'
        404:
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /pages/{pageId}/content:
    parameters:
      - $ref: '#/components/parameters/pageIdParam'
    get:
      tags:
        - Portal
      summary: Get the content of a portal page.
      description: |
        Get the contentn of a specific portal documentation page.
      operationId: getPageContentByPageId
      security: []
      responses:
        200:
          description: The content of the portal documentation page.
          content:
            text/plain:
              schema:
                type: string
        404:
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    apiIdParam:
      name: apiId
      in: path
      required: true
      description: Id of an API.
      schema:
        type: string
...
    pageIdParam:
      name: pageId
      in: path
      required: true
      description: Id of a documentation page.
      schema:
        type: string
...
    pageIncludeParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - content
    #Analytics & logs filtering
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    Error:
      properties:
        status:
          description: Status HTTP of the error
          type: string
        message:
          description: Message of the error
          type: string
        code:
          description: Technical code of the error
          type: string
        parameters:
          description: Parameters
          type: object
          additionalProperties:
            type: string
...
    Page:
      required:
        - id
        - name
        - type
        - order
      properties:
        id:
          description: Unique identifier of a page.
          type: string
        name:
          description: Name of the page.
          type: string
        type:
          description: Type of documentation.
          type: string
          enum:
            - ASCIIDOC
            - SWAGGER
            - MARKDOWN
            - FOLDER
            - ROOT
            - LINK
        order:
          description: Order of the documentation page in its folder.
          type: integer
        parent:
          description: Parent page. MAY be null.
          type: string
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        configuration:
          $ref: '#/components/schemas/PageConfiguration'
        media:
          description: list of media hash, attached to this page
          type: array
          items:
            $ref: '#/components/schemas/PageMedia'
        metadata:
          description: Array of metadata about the page. This array is filled when the page has been fetched from a distant source (GitHub, GitLab, etc...).
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
        _links:
          $ref: '#/components/schemas/PageLinks'

        ###################
        # may be included #
        ###################
        content:
          description: |
            Only returned with */apis/{apiId}/pages/{pageId}* and */pages/{pageId}*.
            Need *include* query param to contain 'content'.

            The content of the page.
          type: string
        contentRevisionId:
          description:
            The page revision used to fill the content.
          $ref: '#/components/schemas/PageRevisionId'
    PageMedia:
      required:
        - name
        - link
        - type
      properties:
        name:
          description: the name of the media.
          type: string
        link:
          description: link to download the media.
          type: string
        type:
          description: type of the media.
          type: string
...
    Metadata:
      description: Information about the page, if this page is from an external source.
      properties:
        name:
          type: string
        value:
          type: string
        order:
          type: string
...
    PageConfiguration:
      description: Technical information about the page
      properties:
        try_it:
          type: boolean
          description: Enable "Try It!" mode in documentation page.
        try_it_anonymous:
          type: boolean
          description: Enable "Try It!" mode in documentation page for anonymous users.
        try_it_url:
          type: string
          description: Base URL used to try the API.
        show_url:
          type: string
          description: Show the URL to download the content.
        display_operation_id:
          type: boolean
          description: Display the operationId in the operations list.
        doc_expansion:
          type: string
          description: |
            Default expansion setting for the operations and tags.\
            Possibles values are :
             - list : Expands only the tags
             - full : Expands the tags and operations
             - none : Expands nothing. DEFAULT.
          enum:
            - list
            - full
            - none
        enable_filtering:
          type: boolean
          description: Add a top bar to filter content.
        show_extensions:
          type: boolean
          description: Display vendor extension (X-) fields and values for Operations, Parameters, and Schema.
        show_common_extensions:
          type: boolean
          description: Display extensions (pattern, maxLength, minLength, maximum, minimum) fields and values for Parameters.
        max_displayed_tags:
          type: number
          description: |
            Number of max tagged operations displayed. \
            Limits the number of tagged operations displayed to at most this many (negative means show all operations).\
            No limit by default.
        viewer:
          type: string
          description: The type of viewer for OpenAPI specification. Default is 'Swagger'
          enum:
            - Swagger
            - Redoc
...
    PageRevisionId:
      properties:
        pageId:
          type: string
        revision:
          type: integer
          format: int32
...
    PageLinks:
      properties:
        self:
          type: string
        content:
          type: string
        parent:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PageSuccess:
      description: One documentation page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Page'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM